+++ src/shared.ts
@@ -1,6 +1,7 @@
 "use strict";
 
 import * as vscode from 'vscode';
+import * as ClientOAuth2 from 'client-oauth2';
 
 export enum UserStatus {
     SignedIn = 1,
@@ -29,7 +30,18 @@ export function showError(message?: string): void {
     vscode.window.showErrorMessage(message);
 }
 
-export const Oauth = {
-    clientID: 'luogu-dev',
-    key: 'HasHidden'
+export const OAuthInfo = {
+    clientID: 'luogu-vscode',
+    key: 'HasHidden',
+    accessTokenUri: 'https://luogu.org/api/OAuth2/accessToken',
+    authorizationUri: 'https://luogu.org/api/OAuth2/authorize',
+    scopes: []
 };
+
+export const luoguAuth = new ClientOAuth2({
+    clientId: OAuthInfo.clientID,
+    clientSecret: OAuthInfo.key,
+    accessTokenUri: OAuthInfo.accessTokenUri,
+    authorizationUri: OAuthInfo.authorizationUri,
+    scopes: OAuthInfo.scopes
+});