+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.0.5
+
+- Add error message when user input is incorrect
+
+- Optimize user expericence
+
 ## 0.0.4
 
 - Optimize user expericence
@@ -13,9 +19,10 @@
 ## 0.0.2
 
 - Fixed bugs
-    1. highlight.js not import
-    2. katex not import
+    1. fix luogu.search
+        1. highlight.js is not imported
+        2. katex is not imported
 
 ## 0.0.1
 
-- Initial release
+- Add searchProblem Function